# repo-governor-contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Repo Governor API
  description: REST API of the Repo Governor SaaS dashboard. All endpoints require an `X-API-Key` header.
  version: 0.1.0
servers:
  - url: http://localhost:8080
security:
  - ApiKeyAuth: []
paths:
...
  /api/v1/organizations/{organizationId}/teams:
    get:
      operationId: listTeams
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
      responses:
        '200':
          description: Teams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Team'
    post:
      operationId: createTeam
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTeamRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    OrganizationId:
      name: organizationId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    Team:
      type: object
      required: [id, organizationId, name]
      properties:
        id:
          type: string
          format: uuid
        organizationId:
          type: string
          format: uuid
        name:
          type: string
    CreateTeamRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
...