# repo-governor-server/src/main/java/io/repogovernor/server/team/TeamController.java
package io.repogovernor.server.team;

import io.repogovernor.server.team.TeamService.TeamDto;
import jakarta.validation.Valid;
import jakarta.validation.constraints.NotBlank;
import org.springframework.http.HttpStatus;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.ResponseStatus;
import org.springframework.web.bind.annotation.RestController;

import java.util.List;
import java.util.UUID;

@RestController
@RequestMapping("/api/v1/organizations/{organizationId}/teams")
public class TeamController {

    private final TeamService teamService;

    public TeamController(TeamService teamService) {
        this.teamService = teamService;
    }

    @GetMapping
    public List<TeamDto> list(@PathVariable UUID organizationId) {
        return teamService.list(organizationId);
    }

    @PostMapping
    @ResponseStatus(HttpStatus.CREATED)
    public TeamDto create(@PathVariable UUID organizationId, @Valid @RequestBody CreateTeamRequest request) {
        return teamService.create(organizationId, request.name());
    }

    public record CreateTeamRequest(@NotBlank String name) {
    }
}